Updated answer to Q11 under .billboard table
According to the question, group names should not include Madonna or Phil, meaning both of these names should not be present in the group name. 

Using the OR statement includes group names if the one the condition is met (NOT madonna or NOT phil collins). This will include some group names with Madonna or Phil Collins in the results. Making both conditions mandatory by using AND eliminates that problem.